pyproject.toml
[project]
name = "colonoscy-anatomical-location"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyrootutils>=1.0.4",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
